refactor(backups): separate full and incremental VM runners
